refactor(ts): simplify toNgffZarr, fromNgffZarr
